requirements.txt
aioelasticsearch==0.5.2
aiohttp==3.0.9
async-timeout==2.0.1
attrs==17.4.0
baidu-aip==2.2.0.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
elasticsearch==6.1.1
Flask==0.12.2
idna==2.6
idna-ssl==1.0.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
multidict==4.1.0
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
yarl==1.1.1
